move ruby file post process to abstract ruby class
