+++ samples/client/petstore/ruby/lib/petstore/api/fake_api.rb
@@ -724,5 +724,6 @@ module Petstore
       end
       return data, status_code, headers
     end
+
   end
 end
