+++ samples/client/petstore/ruby/lib/petstore/api/fake_classname_tags123_api.rb
@@ -73,5 +73,6 @@ module Petstore
       end
       return data, status_code, headers
     end
+
   end
 end
