+++ samples/client/petstore/ruby/lib/petstore/api/pet_api.rb
@@ -508,5 +508,6 @@ module Petstore
       end
       return data, status_code, headers
     end
+
   end
 end
