+++ samples/client/petstore/ruby/lib/petstore/api/store_api.rb
@@ -227,5 +227,6 @@ module Petstore
       end
       return data, status_code, headers
     end
+
   end
 end
