+++ samples/client/petstore/ruby/lib/petstore/api/user_api.rb
@@ -422,5 +422,6 @@ module Petstore
       end
       return data, status_code, headers
     end
+
   end
 end
